Reduce callback query and implement request lookup when file uploaded without replying

+++ src/main/java/com/pirasalbe/services/telegram/handlers/AbstractTelegramHandlerService.java
@@ -123,7 +123,7 @@ public class AbstractTelegramHandlerService {
 		SendResponse response = bot.execute(sendMessage);
 
 		// schedule delete
-		if (delete) {
+		if (delete && response.isOk()) {
 			schedulerService.schedule(
 					(b, r) -> b.execute(new DeleteMessage(r.message().chat().id(), r.message().messageId())), response,
 					timeout, timeUnit);
